calculator/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pyswisseph==2.10.3.2
pydantic==2.7.0
pytz==2024.1
